Fixed final comments from review
