+++ package.json
@@ -14,7 +14,6 @@
   },
   "dependencies": {
     "buffer": "^6.0.3",
-    "delay": "^5.0.0",
     "moment": "^2.29.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
