--- src/components/AgentAuthority/LicenseCredDef/index.js
@@ -1,1 +1,0 @@
-export { default } from './LicenseCredDef'
